refactor: theme context into provider and hook
