+++ src/contexts/ThemeProvider.tsx
@@ -1,0 +1,22 @@
+import { useState } from 'react';
+import { ReactNode } from 'react';
+
+import { ThemeContext } from './ThemeContext';
+
+interface ThemeProviderProps {
+  children: ReactNode;
+}
+
+export const ThemeProvider = ({ children }: ThemeProviderProps) => {
+  const [theme, setTheme] = useState<'light' | 'dark'>('dark');
+
+  const toggleTheme = () => {
+    setTheme((prevTheme) => (prevTheme === 'dark' ? 'light' : 'dark'));
+  };
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      <div data-theme={theme}>{children}</div>
+    </ThemeContext.Provider>
+  );
+};
